+++ route_verification/src/evcxr_examples/as_stats.rs
@@ -168,7 +168,7 @@ fn gen_up_down_hill_stats(query: QueryIr, mut bgp_lines: Vec<Line>, db: AsRelDb)
 /// Copy this after running code from [`parse_bgp_lines`],
 fn gen_as_stats(query: QueryIr, mut bgp_lines: Vec<Line>, db: AsRelDb) -> Result<()> {
     let start = Instant::now();
-    let map: DashMap<u32, RouteStats<u32>> = DashMap::new();
+    let map: DashMap<u32, RouteStats<u64>> = DashMap::new();
     bgp_lines.par_iter_mut().for_each(|l| {
         l.compare.as_stats(&query, &db, &map);
     });
